+++ src/JestExt.ts
@@ -501,7 +501,7 @@ export class JestExt {
 
           case '#!/bin/sh': {
             const line = lines[8]
-            const match = /^\s*"[^"]+"\s+"$basedir\/([^"]+)"/.exec(line)
+            const match = /^\s*"[^"]+"\s+"\$basedir\/([^"]+)"/.exec(line)
             if (match) {
               return path.join(path.dirname(jest), match[1])
             }
@@ -511,6 +511,12 @@ export class JestExt {
         }
 
         break
+      }
+
+      case 'npm test --':
+      case 'npm.cmd test --': {
+        vscode.window.showErrorMessage('Debugging of tasks is currently only available when directly running jest!')
+        return undefined
       }
     }